pyproject.toml
[project]
name = "copier"
version = "1.0.0"
description = "Programmatically copy file to and from configured locations"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "argparse>=1.4.0",
    "pyyaml>=6.0.3",
]

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
    "Environment :: Console",
]
authors = [
    { name = "Alvise Bruniera", email = "alvisebruniera@gmail.com" },
    { name = "Lucide", email = "33134209+Lucide@users.noreply.github.com" },
]

[project.scripts]
copier = "copier:main"

[project.urls]
Repository = "https://github.com/KayJay7/copier.git"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
